Core Plotting Function Collection

Contains core functions for integrating graphic elements and generating the final ggplot2 object.

Integrates all graphic elements to generate the final chord plot

chordPlotFunc(
  allRibbon,
  ribbon_alpha,
  ribbon_color_scheme,
  ribbon_colors,
  show_legend,
  gene_polys,
  gene_pal,
  gene_color_scheme,
  final_gene_order,
  seqArcs,
  gene_arrows,
  gene_label_show,
  gene_label_size,
  show_axis,
  axisLines,
  axisTicks,
  axisLabelOrientation,
  seq_colors,
  seq_labels,
  seqs,
  extremes,
  panel_margin,
  title
)

Arguments

allRibbon

Ribbon data (data.frame)

ribbon_alpha

Ribbon transparency

ribbon_color_scheme

Ribbon color scheme

ribbon_colors

Ribbon color parameters

show_legend

Whether to display the legend

gene_polys

Gene arrow polygon data

gene_pal

Gene color vector

gene_color_scheme

Gene color scheme

final_gene_order

Order of genes in the legend

seqArcs

Sequence arc data

gene_arrows

Gene label data

gene_label_show

Whether to display gene labels

gene_label_size

Font size of gene labels

show_axis

Whether to display axes

axisLines

Axis line data

axisTicks

Tick mark data

axisLabelOrientation

Orientation of axis labels

seq_colors

Sequence color vector

seq_labels

Sequence labels

seqs

List of sequence IDs

extremes

Plot extrema (min/max coordinates)

panel_margin

Panel margins

title

Plot title

process_panel_margin

Function to process panel margins (internal)

Value

ggplot2 object