Add an alignment ribbon layer

Draws colored ribbons corresponding to alignment results. Color scheme and spacing parameters are specified here.

Usage

geom_ribbon(
  mapping = NULL,
  data = NULL,
  ribbon_color_scheme = NULL,
  ribbon_colors = NULL,
  ribbon_color_by = NULL,
  ribbon_color_limits = NULL,
  ribbon_color_breaks = NULL,
  ribbon_color_name = NULL,
  ribbon_alpha = NULL,
  ribbon_alpha_by = NULL,
  ribbon_alpha_range = c(0.15, 0.9),
  ribbon_ctrl_point = NULL,
  ribbon_gap = NULL,
  alpha = NULL,
  ribbon_outline_color = "black",
  ribbon_outline_width = 0.05,
  ribbon_outline_linetype = 1,
  ribbon_outline_by = NULL,
  ribbon_outline_colors = NULL,
  ribbon_linetype_by = NULL,
  ribbon_linetypes = NULL,
  ribbon_direction = c("none", "alpha", "outline", "linetype"),
  ribbon_direction_colors = c(same = "black", reverse = "grey50"),
  ribbon_direction_linetypes = c(same = "solid", reverse = "dashed"),
  ribbon_direction_alpha = c(same = 1, reverse = 0.45),
  show_legend = TRUE,
  legend_position = "left",
  legend_key_length = NULL,
  ...
)

Arguments

mapping

Default NULL (uses pre-computed data)

data

Default NULL (retrieved automatically from the layout)

ribbon_color_scheme

Character. Color scheme "pident", "query", "subject" or "single", default "pident"

ribbon_colors

Optional color vector. Ribbon color parameters

ribbon_color_by

Optional character column name. When set, ribbon fill is mapped to a continuous colourbar for that numeric column instead of pident (e.g. "bitscore").

ribbon_color_limits

Optional numeric length-2 limits for ribbon_color_by.

ribbon_color_breaks

Optional numeric breaks for the ribbon_color_by colourbar.

ribbon_color_name

Optional legend title for the ribbon_color_by colourbar (defaults to the column name).

ribbon_alpha

Numeric (0-1). Ribbon transparency, default 0.35

ribbon_alpha_by

Optional character column name. When set, alpha is scaled continuously from that numeric column.

ribbon_alpha_range

Numeric length-2. Alpha range used by ribbon_alpha_by, default c(0.15, 0.9).

ribbon_ctrl_point

Optional vector/list. Bezier control points, default c(0,0)

ribbon_gap

Optional numeric/vector. Spacing between sequences and ribbons, default 0.15

alpha

Ribbon transparency (overrides ribbon_alpha), defaults to the value used in the layout

ribbon_outline_color

Character. Color of the ribbon outline (border), default "black"

ribbon_outline_width

Numeric. Line width of the ribbon outline, default 0.05

ribbon_outline_linetype

Numeric or character. Line type of the ribbon outline, default 1 (solid); see linetype in ggplot2 for options

ribbon_outline_by

Optional discrete column name. When set, outline colour is mapped by that column and ribbon_outline_colors controls the palette.

ribbon_outline_colors

Optional named color vector for ribbon_outline_by; unnamed vectors are recycled positionally.

ribbon_linetype_by

Optional discrete column name. When set, outline linetype is mapped by that column and ribbon_linetypes controls the values.

ribbon_linetypes

Optional named linetype vector for ribbon_linetype_by.

ribbon_direction

Character. How to visually distinguish same- vs reverse-orientation alignments: "none", "alpha", "outline" or "linetype".

ribbon_direction_colors

Named color vector with same and reverse entries, used when ribbon_direction = "outline".

ribbon_direction_linetypes

Named linetype vector with same and reverse entries, used when ribbon_direction = "linetype".

ribbon_direction_alpha

Named numeric vector with same and reverse entries, used when ribbon_direction = "alpha".

show_legend

Whether to show the legend, default TRUE

legend_position

Position of this layer's legend (the Identity( colourbar): one of "left", "right", "top", "bottom" or "inside", default "left". Pass NULL to let the legend follow theme(legend.position = ...) together with the other legends.

legend_key_length

Optional length of the Identity( long dimension: its height when placed on the left/right, its width when placed on the top/bottom). Accepts a grid unit, e.g. unit(5, "cm"), or a number interpreted as centimetres. Default NULL lets the vertical bar fill the available height (and the horizontal bar default to 4 cm).

...

Additional arguments passed to geom_polygon()

Value

A list of ggplot2 layers

Examples

data(seq_data_example)
data(ribbon_data_example)
p <- ggchord(seq_data_example, ribbon_data_example) +
  geom_seq() + geom_ribbon()
ggplot2::ggplot_build(p)
#> <ggplot2::ggplot_built>
#>  @ data  :List of 2
#>  .. $ :'data.frame':	2000 obs. of  8 variables:
#>  ..  ..$ x        : num [1:2000] 0.707 0.705 0.703 0.701 0.699 ...
#>  ..  ..$ y        : num [1:2000] 0.707 0.709 0.711 0.713 0.716 ...
#>  ..  ..$ group    : int [1:2000] 1 1 1 1 1 1 1 1 1 1 ...
#>  ..  .. ..- attr(*, "n")= int 4
#>  ..  ..$ colour   : chr [1:2000] "#E41A1C" "#E41A1C" "#E41A1C" "#E41A1C" ...
#>  ..  ..$ PANEL    : Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
#>  ..  ..$ linewidth: num [1:2000] 1.2 1.2 1.2 1.2 1.2 1.2 1.2 1.2 1.2 1.2 ...
#>  ..  ..$ linetype : int [1:2000] 1 1 1 1 1 1 1 1 1 1 ...
#>  ..  ..$ alpha    : logi [1:2000] NA NA NA NA NA NA ...
#>  .. $ :'data.frame':	6200 obs. of  9 variables:
#>  ..  ..$ x        : num [1:6200] 0.145 0.135 0.125 0.115 0.106 ...
#>  ..  ..$ y        : num [1:6200] 0.837 0.839 0.841 0.842 0.843 ...
#>  ..  ..$ group    : int [1:6200] 1 1 1 1 1 1 1 1 1 1 ...
#>  ..  .. ..- attr(*, "n")= int 31
#>  ..  ..$ alpha    : num [1:6200] 0.35 0.35 0.35 0.35 0.35 0.35 0.35 0.35 0.35 0.35 ...
#>  ..  ..$ zfill    : chr [1:6200] "#EDE527" "#EDE527" "#EDE527" "#EDE527" ...
#>  ..  ..$ PANEL    : Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
#>  ..  ..$ colour   : chr [1:6200] "black" "black" "black" "black" ...
#>  ..  ..$ linewidth: num [1:6200] 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 0.05 ...
#>  ..  ..$ linetype : num [1:6200] 1 1 1 1 1 1 1 1 1 1 ...
#>  @ layout:Classes 'Layout', 'ggproto', 'gg' <ggproto object: Class Layout, gg>
#>     coord: <ggproto object: Class CoordCartesian, Coord, gg>
#>         aspect: function
#>         backtransform_range: function
#>         clip: off
#>         default: FALSE
#>         distance: function
#>         draw_panel: function
#>         expand: TRUE
#>         is_free: function
#>         is_linear: function
#>         labels: function
#>         limits: list
#>         modify_scales: function
#>         range: function
#>         ratio: 1
#>         render_axis_h: function
#>         render_axis_v: function
#>         render_bg: function
#>         render_fg: function
#>         reverse: none
#>         setup_data: function
#>         setup_layout: function
#>         setup_panel_guides: function
#>         setup_panel_params: function
#>         setup_params: function
#>         train_panel_guides: function
#>         transform: function
#>         super:  <ggproto object: Class CoordCartesian, Coord, gg>
#>     coord_params: list
#>     facet: <ggproto object: Class FacetNull, Facet, gg>
#>         attach_axes: function
#>         attach_strips: function
#>         compute_layout: function
#>         draw_back: function
#>         draw_front: function
#>         draw_labels: function
#>         draw_panel_content: function
#>         draw_panels: function
#>         finish_data: function
#>         format_strip_labels: function
#>         init_gtable: function
#>         init_scales: function
#>         map_data: function
#>         params: list
#>         set_panel_size: function
#>         setup_data: function
#>         setup_panel_params: function
#>         setup_params: function
#>         shrink: TRUE
#>         train_scales: function
#>         vars: function
#>         super:  <ggproto object: Class FacetNull, Facet, gg>
#>     facet_params: list
#>     finish_data: function
#>     get_scales: function
#>     layout: data.frame
#>     map_position: function
#>     panel_params: list
#>     panel_scales_x: list
#>     panel_scales_y: list
#>     render: function
#>     render_labels: function
#>     reset_scales: function
#>     resolve_label: function
#>     setup: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     train_position: function
#>     super:  <ggproto object: Class Layout, gg> 
#>  @ plot  : <ggplot2::ggplot>
#>  .. @ data       : list()
#>  .. .. - attr(*, "class")= chr "waiver"
#>  .. @ layers     :List of 2
#>  .. .. $ geom_seq   :Classes 'LayerInstance', 'Layer', 'ggproto', 'gg' <ggproto object: Class LayerInstance, Layer, gg>
#>     aes_params: list
#>     compute_aesthetics: function
#>     compute_geom_1: function
#>     compute_geom_2: function
#>     compute_position: function
#>     compute_statistic: function
#>     computed_geom_params: list
#>     computed_mapping: ggplot2::mapping, uneval, gg, S7_object
#>     computed_stat_params: list
#>     constructor: call
#>     data: data.frame
#>     draw_geom: function
#>     finish_statistics: function
#>     geom: <ggproto object: Class GeomPath, Geom, gg>
#>         aesthetics: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         draw_group: function
#>         draw_key: function
#>         draw_layer: function
#>         draw_panel: function
#>         extra_params: na.rm
#>         handle_na: function
#>         non_missing_aes: linewidth colour linetype
#>         optional_aes: 
#>         parameters: function
#>         rename_size: TRUE
#>         required_aes: x y
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class GeomPath, Geom, gg>
#>     geom_params: list
#>     ggchord_params: list
#>     ggchord_placeholder: data.frame
#>     ggchord_ref: environment
#>     ggchord_type: seq
#>     inherit.aes: FALSE
#>     layer_data: function
#>     layout: NULL
#>     map_statistic: function
#>     mapping: ggplot2::mapping, uneval, gg, S7_object
#>     name: NULL
#>     position: <ggproto object: Class PositionIdentity, Position, gg>
#>         aesthetics: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         required_aes: 
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class Position, gg>
#>     print: function
#>     setup_layer: function
#>     show.legend: TRUE FALSE
#>     stat: <ggproto object: Class StatIdentity, Stat, gg>
#>         aesthetics: function
#>         compute_group: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         dropped_aes: 
#>         extra_params: na.rm
#>         finish_layer: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         required_aes: 
#>         retransform: TRUE
#>         setup_data: function
#>         setup_params: function
#>         super:  <ggproto object: Class Stat, gg>
#>     stat_params: list
#>     super:  <ggproto object: Class Layer, gg> 
#>  .. .. $ geom_ribbon:Classes 'LayerInstance', 'Layer', 'ggproto', 'gg' <ggproto object: Class LayerInstance, Layer, gg>
#>     aes_params: list
#>     compute_aesthetics: function
#>     compute_geom_1: function
#>     compute_geom_2: function
#>     compute_position: function
#>     compute_statistic: function
#>     computed_geom_params: list
#>     computed_mapping: ggplot2::mapping, uneval, gg, S7_object
#>     computed_stat_params: list
#>     constructor: call
#>     data: data.frame
#>     draw_geom: function
#>     finish_statistics: function
#>     geom: <ggproto object: Class GeomChordPolygon, GeomPolygon, Geom, gg>
#>         aesthetics: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         draw_group: function
#>         draw_key: function
#>         draw_layer: function
#>         draw_panel: function
#>         extra_params: na.rm
#>         handle_na: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         rename_size: TRUE
#>         required_aes: x y
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class GeomChordPolygon, GeomPolygon, Geom, gg>
#>     geom_params: list
#>     ggchord_params: list
#>     ggchord_placeholder: data.frame
#>     ggchord_ref: environment
#>     ggchord_type: ribbon
#>     inherit.aes: FALSE
#>     layer_data: function
#>     layout: NULL
#>     map_statistic: function
#>     mapping: ggplot2::mapping, uneval, gg, S7_object
#>     name: NULL
#>     position: <ggproto object: Class PositionIdentity, Position, gg>
#>         aesthetics: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         required_aes: 
#>         setup_data: function
#>         setup_params: function
#>         use_defaults: function
#>         super:  <ggproto object: Class Position, gg>
#>     print: function
#>     setup_layer: function
#>     show.legend: TRUE FALSE FALSE
#>     stat: <ggproto object: Class StatIdentity, Stat, gg>
#>         aesthetics: function
#>         compute_group: function
#>         compute_layer: function
#>         compute_panel: function
#>         default_aes: ggplot2::mapping, uneval, gg, S7_object
#>         dropped_aes: 
#>         extra_params: na.rm
#>         finish_layer: function
#>         non_missing_aes: 
#>         optional_aes: 
#>         parameters: function
#>         required_aes: 
#>         retransform: TRUE
#>         setup_data: function
#>         setup_params: function
#>         super:  <ggproto object: Class Stat, gg>
#>     stat_params: list
#>     super:  <ggproto object: Class Layer, gg> 
#>  .. @ scales     :Classes 'ScalesList', 'ggproto', 'gg' <ggproto object: Class ScalesList, gg>
#>     add: function
#>     add_defaults: function
#>     add_missing: function
#>     backtransform_df: function
#>     clone: function
#>     find: function
#>     get_scales: function
#>     has_scale: function
#>     input: function
#>     map_df: function
#>     n: function
#>     non_position_scales: function
#>     scales: list
#>     set_palettes: function
#>     train_df: function
#>     transform_df: function
#>     super:  <ggproto object: Class ScalesList, gg> 
#>  .. @ guides     :Classes 'Guides', 'ggproto', 'gg' <ggproto object: Class Guides, gg>
#>     add: function
#>     aesthetics: colour zfill
#>     assemble: function
#>     build: function
#>     draw: function
#>     get_custom: function
#>     get_guide: function
#>     get_params: function
#>     get_position: function
#>     guides: list
#>     merge: function
#>     missing: <ggproto object: Class GuideNone, Guide, gg>
#>         add_title: function
#>         arrange_layout: function
#>         assemble_drawing: function
#>         available_aes: any
#>         build_decor: function
#>         build_labels: function
#>         build_ticks: function
#>         build_title: function
#>         draw: function
#>         draw_early_exit: function
#>         elements: list
#>         extract_decor: function
#>         extract_key: function
#>         extract_params: function
#>         get_layer_key: function
#>         hashables: list
#>         measure_grobs: function
#>         merge: function
#>         override_elements: function
#>         params: list
#>         process_layers: function
#>         setup_elements: function
#>         setup_params: function
#>         train: function
#>         transform: function
#>         super:  <ggproto object: Class GuideNone, Guide, gg>
#>     package_box: function
#>     params: list
#>     print: function
#>     process_layers: function
#>     setup: function
#>     subset_guides: function
#>     train: function
#>     update_params: function
#>     super:  <ggproto object: Class Guides, gg> 
#>  .. @ mapping    : <ggplot2::mapping>  Named list()
#>  .. @ theme      : <theme> List of 144
#>  .. .. $ line                            : <ggplot2::element_line>
#>  .. ..  ..@ colour       : chr "black"
#>  .. ..  ..@ linewidth    : num 0.5
#>  .. ..  ..@ linetype     : num 1
#>  .. ..  ..@ lineend      : chr "butt"
#>  .. ..  ..@ linejoin     : chr "round"
#>  .. ..  ..@ arrow        : logi FALSE
#>  .. ..  ..@ arrow.fill   : chr "black"
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ rect                            : <ggplot2::element_rect>
#>  .. ..  ..@ fill         : chr "white"
#>  .. ..  ..@ colour       : chr "black"
#>  .. ..  ..@ linewidth    : num 0.5
#>  .. ..  ..@ linetype     : num 1
#>  .. ..  ..@ linejoin     : chr "round"
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ text                            : <ggplot2::element_text>
#>  .. ..  ..@ family       : chr ""
#>  .. ..  ..@ face         : chr "plain"
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : chr "black"
#>  .. ..  ..@ size         : num 11
#>  .. ..  ..@ hjust        : num 0.5
#>  .. ..  ..@ vjust        : num 0.5
#>  .. ..  ..@ angle        : num 0
#>  .. ..  ..@ lineheight   : num 0.9
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 0
#>  .. ..  ..@ debug        : logi FALSE
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ title                           : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : NULL
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ point                           : <ggplot2::element_point>
#>  .. ..  ..@ colour       : chr "black"
#>  .. ..  ..@ shape        : num 19
#>  .. ..  ..@ size         : num 1.5
#>  .. ..  ..@ fill         : chr "white"
#>  .. ..  ..@ stroke       : num 0.5
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ polygon                         : <ggplot2::element_polygon>
#>  .. ..  ..@ fill         : chr "white"
#>  .. ..  ..@ colour       : chr "black"
#>  .. ..  ..@ linewidth    : num 0.5
#>  .. ..  ..@ linetype     : num 1
#>  .. ..  ..@ linejoin     : chr "round"
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ geom                            : <ggplot2::element_geom>
#>  .. ..  ..@ ink        : chr "black"
#>  .. ..  ..@ paper      : chr "white"
#>  .. ..  ..@ accent     : chr "#3366FF"
#>  .. ..  ..@ linewidth  : num 0.5
#>  .. ..  ..@ borderwidth: num 0.5
#>  .. ..  ..@ linetype   : int 1
#>  .. ..  ..@ bordertype : int 1
#>  .. ..  ..@ family     : chr ""
#>  .. ..  ..@ fontsize   : num 3.87
#>  .. ..  ..@ pointsize  : num 1.5
#>  .. ..  ..@ pointshape : num 19
#>  .. ..  ..@ colour     : NULL
#>  .. ..  ..@ fill       : NULL
#>  .. .. $ spacing                         : 'simpleUnit' num 5.5points
#>  .. ..  ..- attr(*, "unit")= int 8
#>  .. .. $ margins                         : <ggplot2::margin> num [1:4] 5.5 5.5 5.5 5.5
#>  .. .. $ aspect.ratio                    : NULL
#>  .. .. $ axis.title                      : <ggplot2::element_blank>
#>  .. .. $ axis.title.x                    : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 1
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 2.75 0 0 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.title.x.top                : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 0
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 2.75 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.title.x.bottom             : NULL
#>  .. .. $ axis.title.y                    : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 1
#>  .. ..  ..@ angle        : num 90
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.75 0 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.title.y.left               : NULL
#>  .. .. $ axis.title.y.right              : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 1
#>  .. ..  ..@ angle        : num -90
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 2.75
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.text                       : <ggplot2::element_blank>
#>  .. .. $ axis.text.x                     : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 1
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 2.2 0 0 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.text.x.top                 : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : num 0
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 2.2 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.text.x.bottom              : NULL
#>  .. .. $ axis.text.y                     : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : num 1
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.2 0 0
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.text.y.left                : NULL
#>  .. .. $ axis.text.y.right               : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : num 0
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 0 0 2.2
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.text.theta                 : NULL
#>  .. .. $ axis.text.r                     : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : NULL
#>  .. ..  ..@ hjust        : num 0.5
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : <ggplot2::margin> num [1:4] 0 2.2 0 2.2
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi TRUE
#>  .. .. $ axis.ticks                      : <ggplot2::element_blank>
#>  .. .. $ axis.ticks.x                    : NULL
#>  .. .. $ axis.ticks.x.top                : NULL
#>  .. .. $ axis.ticks.x.bottom             : NULL
#>  .. .. $ axis.ticks.y                    : NULL
#>  .. .. $ axis.ticks.y.left               : NULL
#>  .. .. $ axis.ticks.y.right              : NULL
#>  .. .. $ axis.ticks.theta                : NULL
#>  .. .. $ axis.ticks.r                    : NULL
#>  .. .. $ axis.minor.ticks.x.top          : NULL
#>  .. .. $ axis.minor.ticks.x.bottom       : NULL
#>  .. .. $ axis.minor.ticks.y.left         : NULL
#>  .. .. $ axis.minor.ticks.y.right        : NULL
#>  .. .. $ axis.minor.ticks.theta          : NULL
#>  .. .. $ axis.minor.ticks.r              : NULL
#>  .. .. $ axis.ticks.length               : 'rel' num 0.5
#>  .. .. $ axis.ticks.length.x             : NULL
#>  .. .. $ axis.ticks.length.x.top         : NULL
#>  .. .. $ axis.ticks.length.x.bottom      : NULL
#>  .. .. $ axis.ticks.length.y             : NULL
#>  .. .. $ axis.ticks.length.y.left        : NULL
#>  .. .. $ axis.ticks.length.y.right       : NULL
#>  .. .. $ axis.ticks.length.theta         : NULL
#>  .. .. $ axis.ticks.length.r             : NULL
#>  .. .. $ axis.minor.ticks.length         : 'rel' num 0.75
#>  .. .. $ axis.minor.ticks.length.x       : NULL
#>  .. .. $ axis.minor.ticks.length.x.top   : NULL
#>  .. .. $ axis.minor.ticks.length.x.bottom: NULL
#>  .. .. $ axis.minor.ticks.length.y       : NULL
#>  .. .. $ axis.minor.ticks.length.y.left  : NULL
#>  .. .. $ axis.minor.ticks.length.y.right : NULL
#>  .. .. $ axis.minor.ticks.length.theta   : NULL
#>  .. .. $ axis.minor.ticks.length.r       : NULL
#>  .. .. $ axis.line                       : <ggplot2::element_blank>
#>  .. .. $ axis.line.x                     : NULL
#>  .. .. $ axis.line.x.top                 : NULL
#>  .. .. $ axis.line.x.bottom              : NULL
#>  .. .. $ axis.line.y                     : NULL
#>  .. .. $ axis.line.y.left                : NULL
#>  .. .. $ axis.line.y.right               : NULL
#>  .. .. $ axis.line.theta                 : NULL
#>  .. .. $ axis.line.r                     : NULL
#>  .. .. $ legend.background               : <ggplot2::element_blank>
#>  .. .. $ legend.margin                   : NULL
#>  .. .. $ legend.spacing                  : 'simpleUnit' num 5mm
#>  .. ..  ..- attr(*, "unit")= int 7
#>  .. .. $ legend.spacing.x                : NULL
#>  .. .. $ legend.spacing.y                : NULL
#>  .. .. $ legend.key                      : <ggplot2::element_rect>
#>  .. ..  ..@ fill         : logi NA
#>  .. ..  ..@ colour       : logi NA
#>  .. ..  ..@ linewidth    : NULL
#>  .. ..  ..@ linetype     : NULL
#>  .. ..  ..@ linejoin     : NULL
#>  .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. $ legend.key.size                 : 'simpleUnit' num 1.2lines
#>  .. ..  ..- attr(*, "unit")= int 3
#>  .. .. $ legend.key.height               : NULL
#>  .. .. $ legend.key.width                : NULL
#>  .. .. $ legend.key.spacing              : NULL
#>  .. .. $ legend.key.spacing.x            : NULL
#>  .. .. $ legend.key.spacing.y            : NULL
#>  .. .. $ legend.key.justification        : NULL
#>  .. .. $ legend.frame                    : NULL
#>  .. .. $ legend.ticks                    : NULL
#>  .. .. $ legend.ticks.length             : 'rel' num 0.2
#>  .. .. $ legend.axis.line                : NULL
#>  .. .. $ legend.text                     : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : NULL
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : num 8
#>  .. ..  ..@ hjust        : NULL
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : NULL
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. $ legend.text.position            : NULL
#>  .. .. $ legend.title                    : <ggplot2::element_text>
#>  .. ..  ..@ family       : NULL
#>  .. ..  ..@ face         : chr "bold"
#>  .. ..  ..@ italic       : chr NA
#>  .. ..  ..@ fontweight   : num NA
#>  .. ..  ..@ fontwidth    : num NA
#>  .. ..  ..@ colour       : NULL
#>  .. ..  ..@ size         : num 10
#>  .. ..  ..@ hjust        : num 0
#>  .. ..  ..@ vjust        : NULL
#>  .. ..  ..@ angle        : NULL
#>  .. ..  ..@ lineheight   : NULL
#>  .. ..  ..@ margin       : NULL
#>  .. ..  ..@ debug        : NULL
#>  .. ..  ..@ inherit.blank: logi FALSE
#>  .. .. $ legend.title.position           : NULL
#>  .. .. $ legend.position                 : chr "right"
#>  .. .. $ legend.position.inside          : NULL
#>  .. .. $ legend.direction                : NULL
#>  .. .. $ legend.byrow                    : NULL
#>  .. .. $ legend.justification            : chr "center"
#>  .. .. $ legend.justification.top        : NULL
#>  .. .. $ legend.justification.bottom     : NULL
#>  .. .. $ legend.justification.left       : NULL
#>  .. .. $ legend.justification.right      : NULL
#>  .. .. $ legend.justification.inside     : NULL
#>  .. ..  [list output truncated]
#>  .. .. @ complete: logi TRUE
#>  .. .. @ validate: logi TRUE
#>  .. @ coordinates:Classes 'CoordCartesian', 'Coord', 'ggproto', 'gg' <ggproto object: Class CoordCartesian, Coord, gg>
#>     aspect: function
#>     backtransform_range: function
#>     clip: off
#>     default: FALSE
#>     distance: function
#>     draw_panel: function
#>     expand: TRUE
#>     is_free: function
#>     is_linear: function
#>     labels: function
#>     limits: list
#>     modify_scales: function
#>     range: function
#>     ratio: 1
#>     render_axis_h: function
#>     render_axis_v: function
#>     render_bg: function
#>     render_fg: function
#>     reverse: none
#>     setup_data: function
#>     setup_layout: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     setup_params: function
#>     train_panel_guides: function
#>     transform: function
#>     super:  <ggproto object: Class CoordCartesian, Coord, gg> 
#>  .. @ facet      :Classes 'FacetNull', 'Facet', 'ggproto', 'gg' <ggproto object: Class FacetNull, Facet, gg>
#>     attach_axes: function
#>     attach_strips: function
#>     compute_layout: function
#>     draw_back: function
#>     draw_front: function
#>     draw_labels: function
#>     draw_panel_content: function
#>     draw_panels: function
#>     finish_data: function
#>     format_strip_labels: function
#>     init_gtable: function
#>     init_scales: function
#>     map_data: function
#>     params: list
#>     set_panel_size: function
#>     setup_data: function
#>     setup_panel_params: function
#>     setup_params: function
#>     shrink: TRUE
#>     train_scales: function
#>     vars: function
#>     super:  <ggproto object: Class FacetNull, Facet, gg> 
#>  .. @ layout     :Classes 'Layout', 'ggproto', 'gg' <ggproto object: Class Layout, gg>
#>     coord: NULL
#>     coord_params: list
#>     facet: NULL
#>     facet_params: list
#>     finish_data: function
#>     get_scales: function
#>     layout: NULL
#>     map_position: function
#>     panel_params: NULL
#>     panel_scales_x: NULL
#>     panel_scales_y: NULL
#>     render: function
#>     render_labels: function
#>     reset_scales: function
#>     resolve_label: function
#>     setup: function
#>     setup_panel_guides: function
#>     setup_panel_params: function
#>     train_position: function
#>     super:  <ggproto object: Class Layout, gg> 
#>  .. @ labels     : <ggplot2::labels> List of 8
#>  .. .. $ title : NULL
#>  .. .. $ x     : chr "x"
#>  .. .. $ y     : chr "y"
#>  .. .. $ group : chr "seq_id"
#>  .. .. $ colour: chr "seq_id"
#>  .. .. $ alpha : chr "alpha"
#>  .. .. $ zfill : chr "pident"
#>  .. .. $ alt   : chr ""
#>  .. @ meta       :List of 1
#>  .. .. $ ggchord:List of 5
#>  .. ..  ..$ data      :List of 3
#>  .. ..  .. ..$ seq_data   :'data.frame':	4 obs. of  2 variables:
#>  .. ..  .. .. ..$ seq_id: chr [1:4] "MT108731.1" "MT118296.1" "OQ646790.1" "OR222515.1"
#>  .. ..  .. .. ..$ length: int [1:4] 64323 32090 57367 83080
#>  .. ..  .. ..$ ribbon_data:'data.frame':	31 obs. of  17 variables:
#>  .. ..  .. .. ..$ qaccver   : chr [1:31] "MT108731.1" "MT108731.1" "MT108731.1" "MT108731.1" ...
#>  .. ..  .. .. ..$ saccver   : chr [1:31] "MT118296.1" "MT118296.1" "MT118296.1" "MT118296.1" ...
#>  .. ..  .. .. ..$ pident    : num [1:31] 98.6 97 94.2 96.7 91.7 ...
#>  .. ..  .. .. ..$ length    : int [1:31] 24856 4412 464 302 121 122 5628 3464 2977 1913 ...
#>  .. ..  .. .. ..$ mismatches: int [1:31] 314 125 19 10 10 9 1042 705 600 319 ...
#>  .. ..  .. .. ..$ gapopen   : int [1:31] 25 5 3 0 0 2 128 59 74 26 ...
#>  .. ..  .. .. ..$ qstart    : int [1:31] 26298 21513 20691 20376 19726 19759 45278 59607 52911 57595 ...
#>  .. ..  .. .. ..$ qend      : int [1:31] 51139 25922 21146 20677 19846 19877 50839 63033 55829 59494 ...
#>  .. ..  .. .. ..$ sstart    : int [1:31] 7121 2365 1032 403 5 710 27955 42657 35593 40633 ...
#>  .. ..  .. .. ..$ send      : int [1:31] 31959 6772 1495 704 125 830 33476 46083 38538 42531 ...
#>  .. ..  .. .. ..$ evalue    : num [1:31] 0.00 0.00 0.00 5.93e-143 2.46e-42 ...
#>  .. ..  .. .. ..$ bitscore  : num [1:31] 43960 7417 701 503 169 ...
#>  .. ..  .. .. ..$ qcovs     : int [1:31] 47 47 47 47 47 47 40 40 40 40 ...
#>  .. ..  .. .. ..$ qlen      : int [1:31] 64323 64323 64323 64323 64323 64323 64323 64323 64323 64323 ...
#>  .. ..  .. .. ..$ slen      : int [1:31] 32090 32090 32090 32090 32090 32090 57367 57367 57367 57367 ...
#>  .. ..  .. .. ..$ sstrand   : chr [1:31] "plus" "plus" "plus" "plus" ...
#>  .. ..  .. .. ..$ stitle    : chr [1:31] "N/A" "N/A" "N/A" "N/A" ...
#>  .. ..  .. ..$ gene_data  : NULL
#>  .. ..  ..$ global    :List of 5
#>  .. ..  .. ..$ rotation    : num 45
#>  .. ..  .. ..$ panel_margin: num 0
#>  .. ..  .. ..$ show_legend : logi TRUE
#>  .. ..  .. ..$ debug       : logi FALSE
#>  .. ..  .. ..$ validate    : chr "warn"
#>  .. ..  ..$ validation:List of 7
#>  .. ..  .. ..$ valid       : logi TRUE
#>  .. ..  .. ..$ errors      :'data.frame':	0 obs. of  5 variables:
#>  .. ..  .. .. ..$ table   : chr(0) 
#>  .. ..  .. .. ..$ category: chr(0) 
#>  .. ..  .. .. ..$ row     : int(0) 
#>  .. ..  .. .. ..$ column  : chr(0) 
#>  .. ..  .. .. ..$ message : chr(0) 
#>  .. ..  .. ..$ warnings    :'data.frame':	0 obs. of  5 variables:
#>  .. ..  .. .. ..$ table   : chr(0) 
#>  .. ..  .. .. ..$ category: chr(0) 
#>  .. ..  .. .. ..$ row     : int(0) 
#>  .. ..  .. .. ..$ column  : chr(0) 
#>  .. ..  .. .. ..$ message : chr(0) 
#>  .. ..  .. ..$ summary     :'data.frame':	0 obs. of  4 variables:
#>  .. ..  .. .. ..$ table   : chr(0) 
#>  .. ..  .. .. ..$ category: chr(0) 
#>  .. ..  .. .. ..$ severity: chr(0) 
#>  .. ..  .. .. ..$ n       : int(0) 
#>  .. ..  .. ..$ data_summary:List of 13
#>  .. ..  .. .. ..$ n_seq                : int 4
#>  .. ..  .. .. ..$ n_ribbon             : int 31
#>  .. ..  .. .. ..$ n_gene               : int NA
#>  .. ..  .. .. ..$ n_unknown_ribbon_ids : int 0
#>  .. ..  .. .. ..$ n_unknown_gene_ids   : int 0
#>  .. ..  .. .. ..$ n_self_links         : int 0
#>  .. ..  .. .. ..$ n_out_of_range_ribbon: int 0
#>  .. ..  .. .. ..$ n_out_of_range_gene  : int 0
#>  .. ..  .. .. ..$ n_reversed_ribbon    : int 0
#>  .. ..  .. .. ..$ n_reversed_gene      : int 0
#>  .. ..  .. .. ..$ n_duplicates_ribbon  : int 0
#>  .. ..  .. .. ..$ n_duplicates_gene    : int 0
#>  .. ..  .. .. ..$ n_bad_strand         : int 0
#>  .. ..  .. ..$ invalid_rows: list()
#>  .. ..  .. ..$ cleanable   :'data.frame':	0 obs. of  4 variables:
#>  .. ..  .. .. ..$ table     : chr(0) 
#>  .. ..  .. .. ..$ category  : chr(0) 
#>  .. ..  .. .. ..$ rows      : list()
#>  .. ..  .. .. .. ..- attr(*, "class")= chr "AsIs"
#>  .. ..  .. .. ..$ suggestion: chr(0) 
#>  .. ..  .. ..- attr(*, "class")= chr "ggchord_validation"
#>  .. ..  ..$ ref       :<environment: 0xb6b4e24d8> 
#>  .. ..  ..$ layout    : NULL
#>  .. @ plot_env   :<environment: 0xb6ca76d28>